+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ### Bug fixes
 
 * [#3751](https://github.com/bbatsov/rubocop/pull/3751): Avoid crash in `Rails/EnumUniqueness` cop. ([@pocke][])
+* [#3766](https://github.com/bbatsov/rubocop/pull/3766): Avoid crash in `Style/ConditionalAssignment` cop with masgn. ([@pocke][])
 
 ## 0.46.0 (2016-11-30)